Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    float a;
-    scanf("%f",&a);
+    double a;
+    scanf("%1f",&a);
     double b=a;
     printf("You entered: %1f",a);
     return 0;
